Refactor setup views and add shared components for Ballerina installation
